Install az cli into the container for login
Since the first step in interacting with Azure via these modules is "az
login", ensure that az cli is available at container runtime.

microdnf on EL8 can't install an RPM from a URL (for enabling the
Microsoft packages repo) so instead download the repo RPM with curl and
install it locally with rpm -i.

The Dockerfile.ubi previously made an /opt/aws directory that went
unused. Replace it with an /opt/az in order to have somewhere to store
the RPM between build stages.